+++ Common.Core/ProceduralNoise/PerlinNoise.cs
@@ -7,13 +7,13 @@ namespace Common.Core.ProceduralNoise
 	public class PerlinNoise : Noise
 	{
 
-        public PerlinNoise(int seed, double frequency, double amplitude = 1.0) 
+        public PerlinNoise(int seed, float frequency, float amplitude = 1.0f) 
             : base(seed, frequency, amplitude)
 		{
 
         }
 
-        public PerlinNoise(int seed, Vector3d frequency, double amplitude = 1.0)
+        public PerlinNoise(int seed, Vector3f frequency, float amplitude = 1.0f)
             : base(seed, frequency, amplitude)
         {
 
@@ -28,41 +28,41 @@ namespace Common.Core.ProceduralNoise
             return string.Format("[PerlinNoise: Frequency={0}, Amplitude={1}, Offset={2}]", Frequency, Amplitude, Offset);
         }
 
-        public override double Sample1D( double x )
+        public override float Sample1D( float x )
 		{
             x = (x + Offset.x) * Frequency.x;
 
 		    int ix0;
-		    double fx0, fx1;
-		    double s, n0, n1;
+		    float fx0, fx1;
+		    float s, n0, n1;
 		
 		    ix0 = (int)Math.Floor(x);
 		    fx0 = x - ix0;      
-		    fx1 = fx0 - 1.0;
+		    fx1 = fx0 - 1.0f;
 			
 		    s = FADE(fx0);
 		
 		    n0 = Grad(Perm[ix0], fx0);
 		    n1 = Grad(Perm[ix0 + 1], fx1);
 
-            return 0.25 * LERP(s, n0, n1) * Amplitude;
+            return 0.25f * LERP(s, n0, n1) * Amplitude;
 		}
 		
-		public override double Sample2D( double x, double y )
+		public override float Sample2D( float x, float y )
 		{
             x = (x + Offset.x) * Frequency.x;
             y = (y + Offset.y) * Frequency.y;
 
 		    int ix0, iy0;
-		    double fx0, fy0, fx1, fy1, s, t, nx0, nx1, n0, n1;
+		    float fx0, fy0, fx1, fy1, s, t, nx0, nx1, n0, n1;
 		
 			ix0 = (int)Math.Floor(x); 
 			iy0 = (int)Math.Floor(y); 		
 
 		    fx0 = x - ix0;        	
 		    fy0 = y - iy0;        	
-		    fx1 = fx0 - 1.0;
-		    fy1 = fy0 - 1.0;
+		    fx1 = fx0 - 1.0f;
+		    fy1 = fy0 - 1.0f;
 		    
 		    t = FADE( fy0 );
 		    s = FADE( fx0 );
@@ -82,19 +82,19 @@ namespace Common.Core.ProceduralNoise
 
 		    n1 = LERP(t, nx0, nx1);
 
-            return 0.66666 * LERP(s, n0, n1) * Amplitude;
+            return 0.66666f * LERP(s, n0, n1) * Amplitude;
 		}
 		
-		public override double Sample3D( double x, double y, double z )
+		public override float Sample3D( float x, float y, float z )
 		{
             x = (x + Offset.x) * Frequency.x;
             y = (y + Offset.y) * Frequency.y;
             z = (z + Offset.z) * Frequency.z;
 
             int ix0, iy0, iz0;
-		    double fx0, fy0, fz0, fx1, fy1, fz1;
-		    double s, t, r;
-		    double nxy0, nxy1, nx0, nx1, n0, n1;
+		    float fx0, fy0, fz0, fx1, fy1, fz1;
+		    float s, t, r;
+		    float nxy0, nxy1, nx0, nx1, n0, n1;
 		
 			ix0 = (int)Math.Floor(x);
 			iy0 = (int)Math.Floor(y);  
@@ -102,9 +102,9 @@ namespace Common.Core.ProceduralNoise
 		    fx0 = x - ix0; 
 		    fy0 = y - iy0;  
 		    fz0 = z - iz0;
-		    fx1 = fx0 - 1.0;
-		    fy1 = fy0 - 1.0;
-		    fz1 = fz0 - 1.0;
+		    fx1 = fx0 - 1.0f;
+		    fy1 = fy0 - 1.0f;
+		    fz1 = fz0 - 1.0f;
 		    
 		    r = FADE( fz0 );
 		    t = FADE( fy0 );
@@ -130,43 +130,43 @@ namespace Common.Core.ProceduralNoise
 		
 		    n1 = LERP( t, nx0, nx1 );
 
-            return 1.1111 * LERP(s, n0, n1) * Amplitude;
+            return 1.1111f * LERP(s, n0, n1) * Amplitude;
 		}
 
-        private double FADE(double t) { return t * t * t * (t * (t * 6.0 - 15.0) + 10.0); }
+        private float FADE(float t) { return t * t * t * (t * (t * 6.0f - 15.0f) + 10.0f); }
 
-        private double LERP(double t, double a, double b) { return a + t * (b - a); }
+        private float LERP(float t, float a, float b) { return a + t * (b - a); }
 
-        private double Grad(int hash, double x)
+        private float Grad(int hash, float x)
         {
             int h = hash & 15;
-            double grad = 1.0 + (h & 7);    // Gradient value 1.0, 2.0, ..., 8.0
+            float grad = 1.0f + (h & 7);    // Gradient value 1.0, 2.0, ..., 8.0
             if ((h & 8) != 0) grad = -grad; // Set a random sign for the gradient
             return (grad * x);              // Multiply the gradient with the distance
         }
 
-        private double Grad(int hash, double x, double y)
+        private float Grad(int hash, float x, float y)
         {
             int h = hash & 7;           // Convert low 3 bits of hash code
-            double u = h < 4 ? x : y;  // into 8 simple gradient directions,
-            double v = h < 4 ? y : x;  // and compute the dot product with (x,y).
-            return ((h & 1) != 0 ? -u : u) + ((h & 2) != 0 ? -2.0 * v : 2.0 * v);
+            float u = h < 4 ? x : y;  // into 8 simple gradient directions,
+            float v = h < 4 ? y : x;  // and compute the dot product with (x,y).
+            return ((h & 1) != 0 ? -u : u) + ((h & 2) != 0 ? -2.0f * v : 2.0f * v);
         }
 
-        private double Grad(int hash, double x, double y, double z)
+        private float Grad(int hash, float x, float y, float z)
         {
             int h = hash & 15;     // Convert low 4 bits of hash code into 12 simple
-            double u = h < 8 ? x : y; // gradient directions, and compute dot product.
-            double v = h < 4 ? y : h == 12 || h == 14 ? x : z; // Fix repeats at h = 12 to 15
+            float u = h < 8 ? x : y; // gradient directions, and compute dot product.
+            float v = h < 4 ? y : h == 12 || h == 14 ? x : z; // Fix repeats at h = 12 to 15
             return ((h & 1) != 0 ? -u : u) + ((h & 2) != 0 ? -v : v);
         }
 
-        private double Grad(int hash, double x, double y, double z, double t)
+        private float Grad(int hash, float x, float y, float z, float t)
         {
             int h = hash & 31;          // Convert low 5 bits of hash code into 32 simple
-            double u = h < 24 ? x : y; // gradient directions, and compute dot product.
-            double v = h < 16 ? y : z;
-            double w = h < 8 ? z : t;
+            float u = h < 24 ? x : y; // gradient directions, and compute dot product.
+            float v = h < 16 ? y : z;
+            float w = h < 8 ? z : t;
             return ((h & 1) != 0 ? -u : u) + ((h & 2) != 0 ? -v : v) + ((h & 4) != 0 ? -w : w);
         }
